repo clone preparation
